Add blog post with useful papers
This blog post includes a bunch of papers that I've been meaning to email
out to the Bullmore lab group.

I figured I'd write it up as a blog post instead incase it was useful to
others!

I've also added in an author field to the blog template to remind others
to make sure it doesn't default to me in the future.